#4380 - Refactor (MonomerWithIDTAliasesPreview): Optimize the conditions
