+++ packages/ketcher-macromolecules/src/components/shared/IDTAliases/index.tsx
